Added users and schedules partially
I do not have access to the net ids of users without the netid in their
emails since supabase did not store netids. Therefore, about 900 users
are missing.

Next steps:

Tweak the OIT scraper to account for terms 1232 and before-- for some
reason this breaks.

Add the rest of the users and the rest of the schedules.

Migrate the rest of the data.